[Feat] 다른 씬과의 사운드 연동(+볼륨 슬라이더)
초기 볼륨 슬라이더 설정값 현재 씬에 적용되도록 버그 수정해야함
