+++ Assets/PSH/Audio/MainTheme.mixer
@@ -1,5 +1,19 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!244 &-7701636848982093612
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: 75bf623a68dd1bb43986935c76c7fc05
+  m_EffectName: Attenuation
+  m_MixLevel: b8a1054170438b243961d4435eebe933
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
 --- !u!244 &-6297277943949560535
 AudioMixerEffectController:
   m_ObjectHideFlags: 3
@@ -10,6 +24,58 @@ AudioMixerEffectController:
   m_EffectID: ee27aac72135fbe4a8b2b27e9da5bd31
   m_EffectName: Attenuation
   m_MixLevel: 386709f86ee903547abad69733178919
+  m_Parameters: []
+  m_SendTarget: {fileID: 0}
+  m_EnableWetMix: 0
+  m_Bypass: 0
+--- !u!243 &-5385396238309883189
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Boss2SFX
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: 7c5b7e0b5e622d14f9b8f51eda58334d
+  m_Children: []
+  m_Volume: e4dbe8852750b724fa18610de403e64e
+  m_Pitch: 2c22abf14c982a94a9fe1c774baf3d58
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: -2938194082505472764}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!243 &-3962828188468603730
+AudioMixerGroupController:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: Boss1SFX
+  m_AudioMixer: {fileID: 24100000}
+  m_GroupID: f9acc16ebd59a7146bc9d24c4e1caf22
+  m_Children: []
+  m_Volume: db46a7122be61dd45a928728c8f33522
+  m_Pitch: c0e4d020ce396dc46bcb8bd145d85f9d
+  m_Send: 00000000000000000000000000000000
+  m_Effects:
+  - {fileID: -7701636848982093612}
+  m_UserColorIndex: 0
+  m_Mute: 0
+  m_Solo: 0
+  m_BypassEffects: 0
+--- !u!244 &-2938194082505472764
+AudioMixerEffectController:
+  m_ObjectHideFlags: 3
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_Name: 
+  m_EffectID: 3204146fee32d004a8b7b80d10c915b1
+  m_EffectName: Attenuation
+  m_MixLevel: e857bc7b3f3b3ef4084d36b1fe567816
   m_Parameters: []
   m_SendTarget: {fileID: 0}
   m_EnableWetMix: 0
@@ -39,6 +105,8 @@ AudioMixerController:
     - a6e625d92693d9b44b1959c4f6fad749
     - 4fd6c6ea73eeb5047a39f9d238579a53
     - 6b77f1464373ac149aca932943dba084
+    - f9acc16ebd59a7146bc9d24c4e1caf22
+    - 7c5b7e0b5e622d14f9b8f51eda58334d
     name: View
   m_CurrentViewIndex: 0
   m_TargetSnapshot: {fileID: 24500006}
@@ -87,7 +155,7 @@ AudioMixerSnapshotController:
   m_AudioMixer: {fileID: 24100000}
   m_SnapshotID: 5be128bb78293974faf713471d8e6e98
   m_FloatValues:
-    05431985da324cd4198b3bff4896e450: 3.931848
+    05431985da324cd4198b3bff4896e450: 3.5
     98a69aa67cf31204188cc06e372dc1cb: 0.012838672
   m_TransitionOverrides: {}
 --- !u!243 &1655845931190198451
@@ -99,7 +167,9 @@ AudioMixerGroupController:
   m_Name: SFX
   m_AudioMixer: {fileID: 24100000}
   m_GroupID: 6b77f1464373ac149aca932943dba084
-  m_Children: []
+  m_Children:
+  - {fileID: -3962828188468603730}
+  - {fileID: -5385396238309883189}
   m_Volume: 05431985da324cd4198b3bff4896e450
   m_Pitch: 10654123af7ae5648bb37f0641ac5419
   m_Send: 00000000000000000000000000000000
